[refactor] Moved screen responsiveness and useIsDesktopOrAbove hook to a new utils component for carousel and integrated it into the testimonial component
